Revert "Feature/212"

+++ admission/views/attachments.py
@@ -25,6 +25,7 @@
 ##############################################################################
 from django.shortcuts import render, redirect
 from admission import models as mdl
+from admission.models.applicant_document_file import ApplicantDocumentFile
 from admission.models.enums import document_type
 from admission.views import demande_validation, navigation
 from admission.forms import RemoveAttachmentForm
@@ -34,20 +35,22 @@ from django.forms import formset_factory
 
 
 def update(request, application_id=None):
-    if application_id:
-        application = mdl.application.find_by_id(application_id)
-    else:
-        application = mdl.application.init_application(request.user)
-    past_attachments = list_attachments(application)
+    past_attachments = list_attachments(request.user)
     attachments_available = attachments_left_available(len(past_attachments))
     UploadDocumentFileFormSet = formset_factory(UploadDocumentFileForm, extra=0, max_num=attachments_available)
+
     if request.method == "POST":
         document_formset = UploadDocumentFileFormSet(request.POST, request.FILES)
         if document_formset.is_valid():
             for document in document_formset:
-                save_document_from_form(document, request.user, application)
+                save_document_from_form(document, request.user)
     elif request.method == "GET":
         document_formset = UploadDocumentFileFormSet()
+
+    if application_id:
+        application = mdl.application.find_by_id(application_id)
+    else:
+        application = mdl.application.init_application(request.user)
 
     applicant = mdl.applicant.find_by_user(request.user)
 
@@ -77,31 +80,33 @@ def remove_attachment(request):
             attachment_pk = form.cleaned_data['attachment_id']
             # form is valid ensure that there is a document having that pk value
             attachment_to_remove = DocumentFile.objects.get(pk=attachment_pk)
-            safe_document_removal("admission_attachments", attachment_to_remove)
+            safe_document_removal(request.user, "admission_attachments", attachment_to_remove)
     return redirect(update)
 
 
-def safe_document_removal(application_name, document):
+def safe_document_removal(user, application_name, document):
     """
     Safely remove a document by ensuring that the user is the one
     that owns the file and the application_name is the correct one.
+    :param user: a User object
     :param application_name: a string
     :param document
     :return:
     """
-    if document.application_name == application_name:
+    applicant_calling = mdl.applicant.find_by_user(user)
+    applicant_from_document = mdl.applicant_document_file.find_applicant_by_document(document)
+    if applicant_calling == applicant_from_document and document.application_name == application_name:
         document.delete()
 
 
-def list_attachments(application):
+def list_attachments(user):
     """
     Returns the list of all the attachments uploaded by the user.
-    :param application
+    :param user: the current user in session.
     :return: an array of dictionnary
     """
-    application_document_files = mdl.application_document_file.find_document_by_application(application)
-    return [application_document_file.document_file for application_document_file
-            in application_document_files]
+    applicant = mdl.applicant.find_by_user(user)
+    return mdl.applicant_document_file.find_document_by_applicant(applicant)
 
 
 def attachments_left_available(number_attachments_uploaded):
@@ -115,12 +120,11 @@ def attachments_left_available(number_attachments_uploaded):
     return max_num_attachments - num_attachments_uploaded
 
 
-def save_document_from_form(document, user, application):
+def save_document_from_form(document, user):
     """
     Save a document (attachment) from a form.
     :param document: an UploadDocumentForm received from a POST request.
     :param user: the current user
-    :param application: the current application
     :return:
     """
     file_name = document.cleaned_data['file_name']
@@ -138,7 +142,7 @@ def save_document_from_form(document, user, application):
                             application_name=application_name, content_type=content_type,
                             size=size, update_by=user.username)
     doc_file.save()
-    application_file = mdl.application_document_file.ApplicationDocumentFile()
-    application_file.application = application
-    application_file.document_file = doc_file
-    application_file.save()
+    applicant = mdl.applicant.find_by_user(user)
+    applicant_document_file = ApplicantDocumentFile(applicant=applicant, document_file=doc_file)
+    applicant_document_file.save()
+
